+++ admin/tool/lp/classes/user_evidence_competency.php
@@ -139,6 +139,9 @@ class user_evidence_competency extends persistent {
      */
     public static function delete_by_competencyids($competencyids) {
         global $DB;
+        if (empty($competencyids)) {
+            return true;
+        }
         list($insql, $params) = $DB->get_in_or_equal($competencyids);
         return $DB->delete_records_select(self::TABLE, "competencyid $insql", $params);
     }
